+++ .travis.yml
@@ -1,9 +1,9 @@
 language: node_js
 
 node_js:
-  - 0.8
-  - 0.9
-  - 0.10
+  - "0.8"
+  - "0.9"
+  - "0.10"
 
 script:
   - "npm test"
